# Speed up sampling in the case of measuring all qudits